+++ CHANGELOG.md
@@ -10,6 +10,9 @@
 - Docker image includes the latest versions of dependencies.
 - Optimize processing of animated images.
 
+### Fix
+- Fix error when requested WebP dimension exceeds the WebP dimension limit.
+
 ## [2.13.1] - 2020-05-06
 ### Fixed
 - Fix and optimize processing of animated images.